App_ANN/requirements.txt
absl-py==2.3.1
astunparse==1.6.3
attrs==25.3.0
cachetools==5.5.2
certifi==2025.11.12
cffi==1.17.1
charset-normalizer==3.4.4
comtypes==1.4.12
contourpy==1.1.1
cycler==0.12.1
flatbuffers==25.9.23
fonttools==4.57.0
gast==0.4.0
google-auth==2.43.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.70.0
h5py==3.11.0
idna==3.11
importlib-metadata==8.5.0
importlib-resources==6.4.5
jax==0.4.13
joblib==1.4.2
keras==2.10.0
keras-preprocessing==1.1.2
kiwisolver==1.4.7
libclang==18.1.1
markdown==3.7
markupsafe==2.1.5
matplotlib==3.7.5
mediapipe==0.10.11
ml-dtypes==0.2.0
numpy==1.24.3
oauthlib==3.3.1
opencv-contrib-python==4.12.0.88
opt-einsum==3.4.0
packaging==25.0
pillow==10.4.0
protobuf==3.20.3
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.23
pyparsing==3.1.4
pypiwin32==223
python-dateutil==2.9.0.post0
pywin32==311
requests==2.32.4
requests-oauthlib==2.0.0
rsa==4.9.1
scikit-learn==1.3.2
scipy==1.10.1
six==1.17.0
sounddevice==0.5.3
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-cpu==2.10.0
tensorflow-estimator==2.10.0
tensorflow-intel==2.10.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
threadpoolctl==3.5.0
typing-extensions==4.5.0
urllib3==2.2.3
werkzeug==3.0.6
winspeech==1.0.1
wrapt==2.0.1
zipp==3.20.2
